+++ parl/remote/utils.py
@@ -17,7 +17,8 @@ import os
 from parl.utils import isnotebook
 
 __all__ = [
-    'load_remote_class', 'redirect_stdout_to_file', 'locate_remote_file'
+    'load_remote_class', 'redirect_stdout_to_file', 'locate_remote_file',
+    'get_subfiles_recursively'
 ]
 
 
@@ -137,3 +138,36 @@ def locate_remote_file(module_path):
     else:
         relative_module_path = module_path
     return relative_module_path
+
+
+def get_subfiles_recursively(folder_path):
+    '''
+    Get subfiles under 'folder_path' recursively
+    Args:
+        folder_path: A folder(dir) whose subfiles/subfolders will be returned.
+
+    Returns:
+        python_files: A set including subfiles endwith '.py'.
+        other_files: A set including subfiles not endwith '.py'.
+        empty_subfolders: A set including empty subfolders.
+    '''
+    if not os.path.exists(folder_path):
+        raise ValueError("Path '{}' don't exist.".format(folder_path))
+    elif not os.path.isdir(folder_path):
+        raise ValueError('Input should be a folder, not a file.')
+    else:
+        python_files = set()
+        other_files = set()
+        empty_subfolders = set()
+        for root, dirs, files in os.walk(folder_path):
+            if files:
+                for sub_file in files:
+                    if sub_file.endswith('.py'):
+                        python_files.add(
+                            os.path.normpath(os.path.join(root, sub_file)))
+                    else:
+                        other_files.add(
+                            os.path.normpath(os.path.join(root, sub_file)))
+            elif len(dirs) == 0:
+                empty_subfolders.add(os.path.normpath(root))
+        return python_files, other_files, empty_subfolders
